pyproject.toml
[tool.poetry]
name = "final situational task"
version = "0.1.0"
description = ""
authors = ["Bolot Zhusupekov <bolot.zhusupekov07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
isort = "^5.13.2"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
